Restructure summary modules into mmm/summary package.
Consolidates summary_*.py files into a summary/ subpackage mirroring
plotting/, with __init__.py preserving the public import paths.
