+++ pymc_marketing/mmm/plotting/cv.py
@@ -32,7 +32,7 @@ from pymc_marketing.mmm.plotting._helpers import (
     _process_plot_params,
     _select_dims,
 )
-from pymc_marketing.mmm.summary_cv import (
+from pymc_marketing.mmm.summary.cv import (
     _build_predictions_arrays,
     _crps_for_split,
     _extract_cv_labels,
