+++ pymc_marketing/mmm/plotting/decomposition.py
@@ -34,7 +34,7 @@ from pymc_marketing.mmm.plotting._helpers import (
     _process_plot_params,
     _select_dims,
 )
-from pymc_marketing.mmm.summary_helpers import (
+from pymc_marketing.mmm.summary.helpers import (
     compute_channel_shares,
     compute_waterfall_components,
 )
