+++ pymc_marketing/mmm/summary/__init__.py
@@ -1,0 +1,34 @@
+#   Copyright 2022 - 2026 The PyMC Labs Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+"""Summary DataFrame generation for MMM models."""
+
+from pymc_marketing.data.idata.schema import Frequency
+from pymc_marketing.mmm.summary.budget import BudgetSummaryFactory
+from pymc_marketing.mmm.summary.cv import MMMCVSummaryFactory
+from pymc_marketing.mmm.summary.factory import MMMSummaryFactory
+from pymc_marketing.mmm.summary.helpers import (
+    DataFrameType,
+    OutputFormat,
+    dataframe_to_json_records,
+)
+
+__all__ = [
+    "BudgetSummaryFactory",
+    "DataFrameType",
+    "Frequency",
+    "MMMCVSummaryFactory",
+    "MMMSummaryFactory",
+    "OutputFormat",
+    "dataframe_to_json_records",
+]
